Add support for apache2 --build-all

+++ .github/scripts/build.sh
@@ -1,15 +1,20 @@
+configure_phpbuild() {
+  if [ "$new_version" != "nightly" ]; then
+    sudo cp "$action_dir"/.github/scripts/stable /usr/local/share/php-build/definitions/"$PHP_VERSION"
+    sudo sed -i "s/phpsrctag/$new_version/" /usr/local/share/php-build/definitions/"$PHP_VERSION"
+  else
+    sudo cp "$action_dir"/.github/scripts/"$PHP_VERSION" /usr/local/share/php-build/definitions/
+  fi
+  sudo sed -i "s|PREFIX|$install_dir|" /usr/local/share/php-build/definitions/"$PHP_VERSION"
+}
+
 setup_phpbuild() {
   (
     cd ~ || exit
     git clone git://github.com/php-build/php-build
     cd php-build || exit
     sudo ./install.sh
-    if [ "$new_version" != "nightly" ]; then
-      sudo cp "$action_dir"/.github/scripts/stable /usr/local/share/php-build/definitions/"$PHP_VERSION"
-      sudo sed -i "s/phpsrctag/$new_version/" /usr/local/share/php-build/definitions/"$PHP_VERSION"
-    else
-      sudo cp "$action_dir"/.github/scripts/"$PHP_VERSION" /usr/local/share/php-build/definitions/
-    fi
+    configure_phpbuild
   )
 }
 
@@ -23,10 +28,10 @@ setup_pear() {
   sudo "$install_dir"/bin/pear channel-update pear.php.net
 }
 
-setup_coverage() {
+setup_extensions() {
   sudo "$install_dir"/bin/pecl install -f pcov
-  sudo "$install_dir"/bin/pecl install -f sqlsrv-5.9.0preview1
-  sudo "$install_dir"/bin/pecl install -f pdo_sqlsrv-5.9.0preview1
+  sudo "$install_dir"/bin/pecl install -f sqlsrv-5.9.0beta2
+  sudo "$install_dir"/bin/pecl install -f pdo_sqlsrv-5.9.0beta2
   sudo sed -i "/pcov/d" "$install_dir"/etc/php.ini
   sudo sed -i "/sqlsrv/d" "$install_dir"/etc/php.ini
   sudo chmod a+x .github/scripts/install-ext.sh
@@ -35,6 +40,9 @@ setup_coverage() {
 }
 
 build_php() {
+  export CFLAGS="-Wno-missing-field-initializers $CFLAGS"
+  export CXXFLAGS="-Wno-missing-field-initializers $CXXFLAGS"
+  export PHP_BUILD_APXS="/usr/bin/apxs2"
   if ! php-build -v -i production "$PHP_VERSION" "$install_dir"; then
     echo 'Failed to build PHP'
     exit 1
@@ -46,7 +54,7 @@ build_php() {
     echo "memory_limit=-1"
   ) >>"$install_dir"/etc/php.ini
   setup_pear
-  setup_coverage
+  setup_extensions
   sudo ln -sv "$install_dir"/sbin/php-fpm "$install_dir"/bin/php-fpm
   sudo ln -sf "$install_dir"/bin/* /usr/bin/
   sudo ln -sf "$install_dir"/etc/php.ini /etc/php.ini
@@ -61,6 +69,16 @@ build_php() {
   sudo chmod a+x "$install_dir"/bin/php-fpm-socket-helper
   sudo sed -Ei "s|;pid.*|pid = /run/php/php$PHP_VERSION-fpm.pid|" "$install_dir"/etc/php-fpm.conf
   sudo sed -Ei "s|;error_log.*|error_log = /var/log/php$PHP_VERSION-fpm.log|" "$install_dir"/etc/php-fpm.conf
+  sudo a2dismod php
+  sudo mv /etc/apache2/mods-available/php.load /etc/apache2/mods-available/php"$PHP_VERSION".load
+  sudo cp -fp /etc/apache2/mods-available/php"$PHP_VERSION".load "$install_dir"/etc/apache2/mods-available/
+  sudo cp -fp .github/scripts/apache.conf /etc/apache2/mods-available/php"$PHP_VERSION".conf
+  sudo cp -fp .github/scripts/apache.conf "$install_dir"/etc/apache2/mods-available/php"$PHP_VERSION".conf
+
+  sudo mkdir -p /lib/systemd/system
+  sudo cp -f "$install_dir"/etc/init.d/php-fpm /etc/init.d/php"$PHP_VERSION"-fpm
+  sudo cp -f "$install_dir"/etc/systemd/system/php-fpm.service /lib/systemd/system/php"$PHP_VERSION"-fpm.service
+  sudo service php"$PHP_VERSION"-fpm start
 }
 
 bintray_create_package() {
@@ -88,11 +106,13 @@ build_and_ship() {
     cd "$install_dir"/.. || exit
     sudo XZ_OPT=-e9 tar cfJ php_"$PHP_VERSION"+ubuntu"$release".tar.xz "$PHP_VERSION"
     sudo tar cf - "$PHP_VERSION" | zstd -22 -T0 --ultra > php_"$PHP_VERSION"+ubuntu"$release".tar.zst
-    curl --user "$BINTRAY_USER":"$BINTRAY_KEY" -X DELETE https://api.bintray.com/content/"$BINTRAY_USER"/"$BINTRAY_REPO"/php_"$PHP_VERSION"+ubuntu"$release".tar.xz || true
-    curl --user "$BINTRAY_USER":"$BINTRAY_KEY" -X DELETE https://api.bintray.com/content/"$BINTRAY_USER"/"$BINTRAY_REPO"/php_"$PHP_VERSION"+ubuntu"$release".tar.zst || true
-    curl --user "$BINTRAY_USER":"$BINTRAY_KEY" -T php_"$PHP_VERSION"+ubuntu"$release".tar.xz https://api.bintray.com/content/shivammathur/php/"$PHP_VERSION"-linux/"$PHP_VERSION"+ubuntu"$release"/php_"$PHP_VERSION"+ubuntu"$release".tar.xz || true
-    curl --user "$BINTRAY_USER":"$BINTRAY_KEY" -T php_"$PHP_VERSION"+ubuntu"$release".tar.zst https://api.bintray.com/content/shivammathur/php/"$PHP_VERSION"-linux/"$PHP_VERSION"+ubuntu"$release"/php_"$PHP_VERSION"+ubuntu"$release".tar.zst || true
-    curl --user "$BINTRAY_USER":"$BINTRAY_KEY" -X POST https://api.bintray.com/content/"$BINTRAY_USER"/"$BINTRAY_REPO"/"$PHP_VERSION"-linux/"$PHP_VERSION"+ubuntu"$release"/publish || true
+    if [[ "$GITHUB_MESSAGE" != *no-ship* ]]; then
+      curl --user "$BINTRAY_USER":"$BINTRAY_KEY" -X DELETE https://api.bintray.com/content/"$BINTRAY_USER"/"$BINTRAY_REPO"/php_"$PHP_VERSION"+ubuntu"$release".tar.xz || true
+      curl --user "$BINTRAY_USER":"$BINTRAY_KEY" -X DELETE https://api.bintray.com/content/"$BINTRAY_USER"/"$BINTRAY_REPO"/php_"$PHP_VERSION"+ubuntu"$release".tar.zst || true
+      curl --user "$BINTRAY_USER":"$BINTRAY_KEY" -T php_"$PHP_VERSION"+ubuntu"$release".tar.xz https://api.bintray.com/content/shivammathur/php/"$PHP_VERSION"-linux/"$PHP_VERSION"+ubuntu"$release"/php_"$PHP_VERSION"+ubuntu"$release".tar.xz || true
+      curl --user "$BINTRAY_USER":"$BINTRAY_KEY" -T php_"$PHP_VERSION"+ubuntu"$release".tar.zst https://api.bintray.com/content/shivammathur/php/"$PHP_VERSION"-linux/"$PHP_VERSION"+ubuntu"$release"/php_"$PHP_VERSION"+ubuntu"$release".tar.zst || true
+      curl --user "$BINTRAY_USER":"$BINTRAY_KEY" -X POST https://api.bintray.com/content/"$BINTRAY_USER"/"$BINTRAY_REPO"/"$PHP_VERSION"-linux/"$PHP_VERSION"+ubuntu"$release"/publish || true
+    fi
   )
 }
 
@@ -120,8 +140,8 @@ action_dir=$(pwd)
 tries=10
 existing_version=$(curl -sL https://github.com/shivammathur/php-builder/releases/latest/download/php"$PHP_VERSION".log)
 new_version=$(curl -sL https://api.github.com/repos/php/php-src/git/refs/tags | grep -Po "php-$PHP_VERSION.[0-9]+" | tail -n 1)
-sudo mkdir -p "$install_dir" /usr/local/ssl
-sudo chmod -R 777 /usr/local/php /usr/local/ssl
+sudo mkdir -p "$install_dir" "$install_dir"/"$(apxs -q SYSCONFDIR)"/mods-available /usr/local/ssl
+sudo chmod -R 777 /usr/local/php /usr/local/ssl /usr/include/apache2 /usr/lib/apache2 /etc/apache2/ /var/lib/apache2 /var/log/apache2
 check_stable
 setup_phpbuild
 build_php
